查询导航路径调整　　打印功能
