+++ fronend/imatch/src/pages/details/grade/graresult/index.css
@@ -6,6 +6,7 @@
     height: 50px;
     font-size: 25px;
     line-height:50px; 
+    position: relative;
 }
 .dealData{
     padding:5px;
