+++ fronend/imatch/src/pages/details/grade/graresult/round/index.css
@@ -71,3 +71,8 @@
     text-align: right;
     padding-right: 20px;
 }
+
+.butbox{
+    text-align: right;
+    margin-top: 20px;
+}
